+++ hydromt/data_catalog.py
@@ -1277,10 +1277,10 @@ class DataCatalog(object):
         data_like: Union[str, SourceSpecDict, Path, xr.Dataset, xr.DataArray],
         bbox: Optional[List] = None,
         geom: Optional[gpd.GeoDataFrame] = None,
-        zoom_level: Optional[Union[int, tuple]] = None,
         buffer: Union[float, int] = 0,
         handle_nodata: NoDataStrategy = NoDataStrategy.RAISE,
         align: Optional[bool] = None,
+        zoom_level: Optional[Union[int, tuple]] = None,
         variables: Optional[Union[List, str]] = None,
         time_tuple: Optional[Tuple] = None,
         single_var_as_array: Optional[bool] = True,
