--- hydromt/models/_region/__init__.py
@@ -1,33 +1,0 @@
-from .region import Region
-from .specifyers import (
-    BasinIdRegionSpecifyer,
-    BasinMultipleIdsRegionSpecifyer,
-    BasinPointBboxRegionSpecifyer,
-    BasinPointGeomRegionSpecifyer,
-    BasinPointListRegionSpecifyer,
-    BasinPointRegionSpecifyer,
-    BasinRegionSpecifyer,
-    BboxRegionSpecifyer,
-    DerivedRegionSpecifyer,
-    GeomRegionSpecifyer,
-    GridRegionSpecifyer,
-    MeshRegionSpecifyer,
-    RegionSpecifyer,
-)
-
-__all__ = [
-    "BasinIdRegionSpecifyer",
-    "BasinMultipleIdsRegionSpecifyer",
-    "BasinPointBboxRegionSpecifyer",
-    "BasinPointGeomRegionSpecifyer",
-    "BasinPointListRegionSpecifyer",
-    "BasinPointRegionSpecifyer",
-    "BasinRegionSpecifyer",
-    "BboxRegionSpecifyer",
-    "DerivedRegionSpecifyer",
-    "GeomRegionSpecifyer",
-    "GridRegionSpecifyer",
-    "MeshRegionSpecifyer",
-    "RegionSpecifyer",
-    "Region",
-]
