+++ hydromt/region/__init__.py
@@ -1,0 +1,6 @@
+"""Model for dealing with region specifcation and manipulation."""
+from hydromt.region import Region
+
+__all__ = [
+    "Region",
+]
